+++ Craftsman/Builders/InfrastructureServiceRegistrationBuilder.cs
@@ -23,12 +23,17 @@ public class InfrastructureServiceRegistrationBuilder
         var dbContextClassPath = ClassPathHelper.DbContextClassPath(srcDirectory, "", projectBaseName);
         var utilsClassPath = ClassPathHelper.WebApiResourcesClassPath(srcDirectory, "", projectBaseName);
         var envServiceClassPath = ClassPathHelper.WebApiServicesClassPath(srcDirectory, "", projectBaseName);
+        var hangfireResourceClassPath = ClassPathHelper.HangfireResourcesClassPath(srcDirectory, "", projectBaseName);
+        
         return @$"namespace {classNamespace};
 
 using {dbContextClassPath.ClassNamespace};
 using {utilsClassPath.ClassNamespace};
 using {envServiceClassPath.ClassNamespace};
+using {hangfireResourceClassPath.ClassNamespace};
 using Configurations;
+using Hangfire;
+using Hangfire.MemoryStorage;
 using Microsoft.EntityFrameworkCore;
 
 public static class ServiceRegistration
@@ -37,7 +42,46 @@ public static class ServiceRegistration
     {{
         // DbContext -- Do Not Delete
 
+        services.SetupHangfire(env);
+
         // Auth -- Do Not Delete
+    }}
+}}
+    
+public static class HangfireConfig
+{{
+    public static void SetupHangfire(this IServiceCollection services, IWebHostEnvironment env)
+    {{
+        services.AddScoped<IJobContextAccessor, JobContextAccessor>();
+        services.AddScoped<IJobWithUserContext, JobWithUserContext>();
+        // if you want tags with sql server
+        // var tagOptions = new TagsOptions() {{ TagsListStyle = TagsListStyle.Dropdown }};
+        
+        // var hangfireConfig = new MemoryStorageOptions() {{ }};
+        services.AddHangfire(config =>
+        {{
+            config
+                .SetDataCompatibilityLevel(CompatibilityLevel.Version_170)
+                .UseMemoryStorage()
+                .UseColouredConsoleLogProvider()
+                .UseSimpleAssemblyNameTypeSerializer()
+                .UseRecommendedSerializerSettings()
+                // if you want tags with sql server
+                // .UseTagsWithSql(tagOptions, hangfireConfig)
+                .UseActivator(new JobWithUserContextActivator(services.BuildServiceProvider()
+                    .GetRequiredService<IServiceScopeFactory>()));
+        }});
+        services.AddHangfireServer(options =>
+        {{
+            options.WorkerCount = 10;
+            options.ServerName = $""PeakLims-{{env.EnvironmentName}}"";
+
+            if (Consts.HangfireQueues.List().Length > 0)
+            {{
+                options.Queues = Consts.HangfireQueues.List();
+            }}
+        }});
+
     }}
 }}";
     }