+++ CHANGELOG.md
@@ -12,6 +12,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 * New `GetAll` feature
+* Hangfire integration
+  * ***TBD HOW TO UPDATE***
+
 
 ### Updates